Remove cloud-init requirement with firstboot script
The requirement to run the full cloud-init system inside a VM created from the
centos-7 template has been removed. Instead of cloud-init a firstboot script is
provided and set to run during the first boot of the machine, which it
completely removes itself from (except the log file) after it runs.

Firstboot also has the capability of resizing disks that contain an LVM setup, a
feature that the "growpart" module in cloud-init was lacking.

A small subset of cloud-init configuration modules have been reimplemented:

* Disk growth (runs regardless of configuration)
* SSH key adding
* Hostname setting
* Phoning home