Better handle firstboot logging
Using a log function inside the script was a nice idea for informational
messages, however it fails to capture the stdout/stderr of programs executed
within the script.

The method of logging to the file has been changed to capture all of
stdout/stderr from the script into a log file, configured as part of the
service's ExecStart parameter.